backend/WorkProjectNew/NTKM/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.14
django-cors-headers @ file:///run/media/adminlocal/Ventoy/Python_Frameworks/python/django_cors_headers-4.4.0-py3-none-any.whl#sha256=5c6e3b7fe870876a1efdfeb4f433782c3524078fa0dc9e0195f6706ce7a242f6
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
grid==0.7.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
oauthlib==3.2.2
optional-django==0.3.0
psycopg==3.2.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
python3-openid==3.2.0
PyYAML==6.0.1
react==4.3.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starlette==0.37.2
tldextract==5.1.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zope.event==5.0
zope.interface==6.4.post2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.14
django-cors-headers @ file:///run/media/adminlocal/Ventoy/Python_Frameworks/python/django_cors_headers-4.4.0-py3-none-any.whl#sha256=5c6e3b7fe870876a1efdfeb4f433782c3524078fa0dc9e0195f6706ce7a242f6
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
grid==0.7.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
oauthlib==3.2.2
optional-django==0.3.0
psycopg==3.2.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
python3-openid==3.2.0
PyYAML==6.0.1
react==4.3.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starlette==0.37.2
tldextract==5.1.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zope.event==5.0
zope.interface==6.4.post2
